% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A86_import.R
\name{A86_import}
\alias{A86_import}
\title{Import data and}
\usage{
A86_import(path, lookup)
}
\arguments{
\item{path}{location of the raw data.}

\item{lookup}{an already-imported lookup table dataframe.}
}
\value{
A dataframe.
}
\description{
Import data and
}
