% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggA86.R
\name{ggA86}
\alias{ggA86}
\title{Title}
\usage{
ggA86(
  data,
  x,
  y,
  group,
  colors = c("blue", "red"),
  graph_type = "line",
  expand = c(0, 0),
  limits = c(0, NA),
  error_bar_width = 1,
  xtitle = NULL,
  ytitle = NULL
)
}
\arguments{
\item{data}{a dataframe or object coercible to a datafram to be graph.}

\item{x}{the parameter to be graphed on the x axis.}

\item{y}{the parameter to be graphed on the y axis.}

\item{group}{the grouping variable.}

\item{colors}{the colors to use for the different groups.}

\item{graph_type}{the type of graph to be plotted. Possible arguments are "line" and "bar".}

\item{expand}{the additional buffer added to the vertical limits of the graph.}

\item{limits}{the vertical limits of the graph.}

\item{error_bar_width}{the width of the error bars.}

\item{xtitle}{the title of the x axis}

\item{ytitle}{the title of the y axis}
}
\value{
a ggplot object
}
\description{
Title
}
\examples{

}
