% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_A86.R
\name{theme_A86}
\alias{theme_A86}
\title{Returns the standard ggplot theme for the A86 paper}
\usage{
theme_A86()
}
\value{
A ggplot2 theme.
}
\description{
Returns the standard ggplot theme for the A86 paper
}
\examples{
mtcars2 <- within(mtcars, {
vs <- factor(vs, labels = c("V-shaped", "Straight"))
am <- factor(am, labels = c("Automatic", "Manual"))
cyl  <- factor(cyl)
gear <- factor(gear)
})
p1 <- ggplot2::ggplot(mtcars2) +
ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = gear)) +
ggplot2::labs(
title = "Fuel economy declines as weight increases",
 subtitle = "(1973-74)",
  caption = "Data from the 1974 Motor Trend US magazine.",
   tag = "Figure 1",
   x = "Weight (1000 lbs)",
   y = "Fuel economy (mpg)",
  colour = "Gears"
)
p1 + theme_A86()

}
